[Enhancement #248] Simplify vocabulary exporter tests.
